requirements.txt
tweepy
openai
pinecone-client
python-dotenv
presidio_analyzer
presidio_anonymizer